api/requirements.txt
fastapi==0.115.6
python-binance==1.0.19

